+++ R/0_imports.R
@@ -10,6 +10,6 @@
 #' @importFrom utils combn tail read.table str write.table
 #' @import sdcHierarchies
 #' @importFrom rlang is_scalar_character is_character is_integerish
-#' @importFrom rlang is_scalar_double is_scalar_integerish
+#' @importFrom rlang is_scalar_double is_scalar_integerish is_scalar_logical
 #' @importFrom knitr knit
 NULL