+++ NEWS.md
@@ -4,6 +4,8 @@
 - allow to specify numVarInd also by name via `numVarName` in primarySupression() for dominance rules (`p`, `nk` and `pq`)
 - make use of sampling weights (by replicating) values for dominance rules.
 - don't export unused functionality
+- `createJJFormat()` and `writeJJFormat()` can be used to create text files with
+sdcProblems in "JJs Format"
 
 # sdcTable 0.27
 - use functionality from [`sdcHierarchies`](https://cran.r-project.org/package=sdcHierarchies) to build hierarchies.